requirements.txt
anyio==2.2.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
blis==0.7.4
bravado==11.0.3
bravado-core==5.17.0
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
defusedxml==0.7.0
distlib==0.3.1
entrypoints==0.3
fastapi==0.63.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.0.0/fr_core_news_sm-3.0.0-py3-none-any.whl
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
gunicorn==20.0.4
h11==0.12.0
identify==2.1.0
idna==2.10
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
json5==0.9.5
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==8.7.0
msgpack==1.0.2
murmurhash==1.0.5
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
neptune-client==0.5.1
nest-asyncio==1.5.1
nltk==3.5
nodeenv==1.5.0
notebook==6.2.0
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pathy==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
pluggy==0.13.1
pre-commit==2.10.1
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pyflakes==2.2.0
Pygments==2.8.0
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==5.4.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rfc3987==1.3.8
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
smart-open==3.0.0
smmap==3.0.5
sniffio==1.2.0
spacy==3.0.3
spacy-legacy==3.0.1
srsly==2.4.0
starlette==0.13.6
stop-words==2018.7.23
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
terminado==0.9.2
testpath==0.4.4
thinc==8.0.1
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tox==3.17.1
tqdm==4.58.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
virtualenv==20.4.2
wasabi==0.8.2
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
